requirements.txt
# Backend
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0
redis==5.0.1

# ML & Data Processing
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.3
joblib==1.3.2

# API Clients
requests==2.31.0
aiohttp==3.9.1

# Utilities
python-dateutil==2.8.2
